ParsingEnumns_Assignment
Create an enum for the days of the week.

Prompt the user to enter the current day of the week.

Assign the value to a variable of that enum data type you just created.

Wrap the above statement in a try/catch block and have it print “Please enter an actual day of the week.” to the console if an error occurs.